Replace imageSmoothArc() with Supersampled AA
GD does not provide native anti-aliasing for arcs. Unfortunately,
Ulrich Mierendorff's imageSmoothArc() is bugged and often misses small
blocks of pixels, towards to the end of the arc; this can be observed in
the green slice of the chart produced by example.php. As a result, the
dependency on imageSmoothArc() has been removed in favour of simple,
naïve supersampling.

The supersampling can be optionally disabled for speed, by passing 0 to
PieChartGD->draw(), as it does incur a performance penalty:
	imageSmoothArc(): 63 ms
	super-sampling (4): 114 ms

The above figures compare the amount of time taken to execute
example.png with the previous commit.