Add alpha component to PieChartColor
It is now possible to construct colours with an alpha transparency
component; the following formats are supported:

- ARGB (integer, e.g. '0x7f000000')
- Array (optional 4th index, e.g. '[0, 0, 0, 0.5]')
- RGBA (string, e.g. 'rgba(0, 0, 0, 0.5)')

The new toRGBA() method can be used to output the colour in a
CSS-compatible format string; which is used by many graphics libraries,
such as Imagick.

(Note: I wrote this quite hastily and this branch will definitely need
testing, before it is merged back into master. Much-needed unit tests are
coming soon!)